+++ src/Company/UnauthenticatedCompanyRegistration.php
@@ -7,6 +7,10 @@ declare(strict_types = 1);
 
 namespace Wizaplace\SDK\Company;
 
+/**
+ * Class UnauthenticatedCompanyRegistration
+ * @package Wizaplace\SDK\Company
+ */
 final class UnauthenticatedCompanyRegistration extends CompanyRegistration
 {
     /**
@@ -19,6 +23,14 @@ final class UnauthenticatedCompanyRegistration extends CompanyRegistration
      */
     private $legalRepresentativeLastName;
 
+    /**
+     * UnauthenticatedCompanyRegistration constructor.
+     *
+     * @param string $name
+     * @param string $email
+     * @param string $legalRepresentativeFirstName
+     * @param string $legalRepresentativeLastName
+     */
     public function __construct(string $name, string $email, string $legalRepresentativeFirstName, string $legalRepresentativeLastName)
     {
         parent::__construct($name, $email);
@@ -26,23 +38,43 @@ final class UnauthenticatedCompanyRegistration extends CompanyRegistration
         $this->legalRepresentativeLastName = $legalRepresentativeLastName;
     }
 
+    /**
+     * @return string
+     */
     public function getLegalRepresentativeFirstName(): string
     {
         return $this->legalRepresentativeFirstName;
     }
 
-    public function setLegalRepresentativeFirstName(string $legalRepresentativeFirstName): void
+    /**
+     * @param string $legalRepresentativeFirstName
+     *
+     * @return UnauthenticatedCompanyRegistration
+     */
+    public function setLegalRepresentativeFirstName(string $legalRepresentativeFirstName): self
     {
         $this->legalRepresentativeFirstName = $legalRepresentativeFirstName;
+
+        return $this;
     }
 
+    /**
+     * @return string
+     */
     public function getLegalRepresentativeLastName(): string
     {
         return $this->legalRepresentativeLastName;
     }
 
-    public function setLegalRepresentativeLastName(string $legalRepresentativeLastName): void
+    /**
+     * @param string $legalRepresentativeLastName
+     *
+     * @return UnauthenticatedCompanyRegistration
+     */
+    public function setLegalRepresentativeLastName(string $legalRepresentativeLastName): self
     {
         $this->legalRepresentativeLastName = $legalRepresentativeLastName;
+
+        return $this;
     }
 }