+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.46.1
 
 - Deprecated `\Wizaplace\SDK\Company\CompanyService::registerC2CCompany` use `CompanyService::register` instead
-- Added `\Wizaplace\SDK\Company\CompanyRegistration::isC2C`
+- Added `\Wizaplace\SDK\Company\CompanyC2CRegistration`
 
 ## 1.46.0